chore: move prefixed-variable resolution into variable.rs
resolve_prefixed_variable, resolve_path_prefix, matching_generic_bounds, and
the PathPrefixKind/ResolvedPrefix types move from types.rs to variable.rs, next
to resolve_variable and resolve_unprefixed_variable. The per-kind handlers
(resolve_self_in_*, resolve_method_on_type_prefix, resolve_trait_item_on_prefix,
resolve_bounded_generic_item) stay in types.rs, now pub(super).
